cleanup, as version 0.1.0
1. remove 'game.lua' 'cli.lua', rename 'gui.lua' to 'game.lua'
2. add more details to main menu
3. rename maps
4. update README.md
5. set version to 0.1.0
