+++ maps/demo_cangqiong.lua
