+++ maps/demo_haiyi.lua
