+++ maps/demo_stardust.lua
